#8265 Python : Move append perforation interval to RimWellPath
