+++ ApplicationLibCode/ProjectDataModelCommands/RimcModeledWellPath.h
@@ -44,24 +44,3 @@ private:
     caf::PdmField<double>  m_tieInDepth;
     caf::PdmField<QString> m_lateralName;
 };
-
-//==================================================================================================
-///
-//==================================================================================================
-class RimcModeledWellPath_appendPerforationInterval : public caf::PdmObjectMethod
-{
-    CAF_PDM_HEADER_INIT;
-
-public:
-    RimcModeledWellPath_appendPerforationInterval( caf::PdmObjectHandle* self );
-
-    caf::PdmObjectHandle*            execute() override;
-    bool                             resultIsPersistent() const override;
-    std::unique_ptr<PdmObjectHandle> defaultResult() const override;
-
-private:
-    caf::PdmField<double> m_startMD;
-    caf::PdmField<double> m_endMD;
-    caf::PdmField<double> m_diameter;
-    caf::PdmField<double> m_skinFactor;
-};
